code reorganization: extracted InitialShape into separate files
