+++ src/client/wrap.h
@@ -19,6 +19,7 @@
 
 #pragma once
 
+#include "InitialShape.h"
 #include "PyCallbacks.h"
 #include "logging.h"
 #include "utils.h"
@@ -42,7 +43,6 @@
 #include <iostream>
 #include <iterator>
 #include <map>
-#include <numeric>
 #include <string>
 #include <vector>
 #ifdef _WIN32
@@ -61,47 +61,6 @@ inline void copyToCStr(const std::string& str, char* cstr, size_t& cstrSize) {
 	}
 	cstrSize = str.length() + 1; // returns the actually needed size including terminating null
 }
-
-class InitialShape {
-public:
-	InitialShape(const std::vector<double>& vert);
-	InitialShape(const std::vector<double>& vert, const std::vector<uint32_t>& ind,
-	             const std::vector<uint32_t>& faceCnt);
-	InitialShape(const std::string& path);
-	~InitialShape() {}
-
-	const double* getVertices() const {
-		return mVertices.data();
-	}
-	size_t getVertexCount() const {
-		return mVertices.size();
-	}
-	const uint32_t* getIndices() const {
-		return mIndices.data();
-	}
-	size_t getIndexCount() const {
-		return mIndices.size();
-	}
-	const uint32_t* getFaceCounts() const {
-		return mFaceCounts.data();
-	}
-	size_t getFaceCountsCount() const {
-		return mFaceCounts.size();
-	}
-	const std::string& getPath() const {
-		return mPath;
-	}
-	bool getPathFlag() const {
-		return mPathFlag;
-	}
-
-protected:
-	const std::vector<double> mVertices;
-	std::vector<uint32_t> mIndices;
-	std::vector<uint32_t> mFaceCounts;
-	const std::string mPath;
-	const bool mPathFlag;
-};
 
 class GeneratedModel {
 public:
